app/requirements.txt
anthropic>=0.25.0
msal>=1.28.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
flask>=3.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
